pyproject.toml
[tool.poetry]
name = "backend-internship-app"
version = "0.1.0"
description = ""
authors = ["Hussein Auf <hu.auf2000@gmail.com>"]
readme = "README.md"
packages = [{include = "backend_internship_app"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.2"
django-model-utils = "^4.2.0"
django-extensions = "^3.2.1"
pip = "^22.3"
install = "^1.3.5"
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
django-filter = "^22.1"
djangorestframework-simplejwt = "^5.2.2"
knox = "^0.1.14"
django-rest-knox = "^4.2.0"
authentication = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
